// frontEnd/frontEnd/src/models/IGuc100.ts
export interface IGuc100 {
    id:                           number;
    ProjeNoMan:                   string;
    ProjeSiparisNo:               string;
    TrafoNo:                      string;
    Tarih:                        Date;
    KatBasinaSarim_AG:            number;
    KanalSayisi_AG:               number;
    OrtalamaCevre_AG:             number;
    KatSayisi_AG:                 number;
    IzoleliAxialTelYuksekligi_AG: number;
    IzoleliTelKalinligi_AG:       number;
    AxialTelYuksekligi_AG:        number;
    TelKalinligi_AG:              number;
    BobinAgirligi_AG:             number;
    Paralel:                      number;
    Radyal:                       number;
    CaprazlamaAG:                 number;
    iletkenCinsi_AG:              number;
    Mat_AG:                       number;
    KatBasinaSarim1_YG:           number;
    KanalSayisi_YG1:              number;
    OrtalamaCevre_YG:             number;
    KatSayisi_YG:                 number;
    IzoleliAxialTelYuksekligi_YG: number;
    IzoleliTelKalinligi_YG:       number;
    AxialTelYuksekligi_YG:        number;
    TelKalinligi_YG:              number;
    BobinAgirligi_YG:             number;
    nSayisi_YG:                   number;
    u_Sayisi_YG:                  number;
    CaprazlamaYG1:                number;
    iletkenCinsiYG1:              number;
    mat_YG:                       number;
    cekirdek_tipi:                number;
    A:                            number;
    PencereYuksekligi:            number;
    CekirdekAgirligi:             number;
    CekirdekSaciCinsi:            number;
    CekirdekCapi:                 number;
    GenisCap:                     number;
    DarCap:                       number;
    EnBuyukSacGenisligi_BS:       number;
    KademeSayisi:                 number;
    Step:                         number;
    Guc:                          number;
    DICK:                         number;
    gerilim_tipi:                 number;
    kamaIzo_YG1:                  number;
    Bobin:                        string;
    Silindi:                      number;
    HermetikKazan:                number;
    kazan_cinsi:                  number;
    sogutma_turu:                 number;
    MekanikMuhendisID:            number;
    Kat_Sayisi:                   number;
}